requirements.txt
aiohttp>=3.8.6
aiohttp-basicauth>=1.0.0
dataclasses-json>=0.6.1
Jinja2>=3.1.2
lxml>=4.9.3
lxml-stubs>=0.4.0
netifaces>=0.11.0
packaging>=23.2
prometheus-client>=0.18.0
python-dotenv>=1.0.0
